+++ demos/components-media-list.php
@@ -40,11 +40,11 @@
       </div>
       <pre><code class="language-scss">
 .news-left > li { 
-  @include media(15px child 'figure' 'div'); 
+  @include media-list(15px child 'figure' 'div'); 
   figure { width: 120px; }
 }
 .news-right > li { 
-  @include media(50px child '.media' '.media-body'); 
+  @include media-list(50px child '.media' '.media-body'); 
   figure { width: 120px; }
 }
       </code></pre>